+++ src/main/java/com/example/realestatemanagment/Configs/SecurityConfig.java
@@ -50,12 +50,32 @@ public class SecurityConfig {
             .httpBasic(basic -> basic.disable())
             .cors(Customizer.withDefaults())
             .authorizeHttpRequests(authorize -> authorize
-                            //.requestMatchers("/tenants/**", "/investors/**").authenticated()
+
                             //.anyRequest().permitAll()
             //.authorizeHttpRequests(auth -> auth
 
                     .requestMatchers("/**").permitAll()
-                    //.requestMatchers(HttpMethod.GET,"/maintenances").hasRole("ADMIN")
+                    //.requestMatchers(HttpMethod.GET,"/maintenances").hasRole("INVESTOR")
+                    //.requestMatchers(HttpMethod.POST,"/maintenances").hasRole("INVESTOR")
+                    //.requestMatchers(HttpMethod.PUT,"/maintenances").hasRole("INVESTOR")
+                    //.requestMatchers(HttpMethod.DELETE,"/maintenances").hasRole("INVESTOR")
+                    //.requestMatchers(HttpMethod.GET,"/properties").hasRole("INVESTOR")
+                    //.requestMatchers(HttpMethod.POST,"/properties").hasRole("INVESTOR")
+                    //.requestMatchers(HttpMethod.PUT,"/properties").hasRole("INVESTOR")
+                    //.requestMatchers(HttpMethod.DELETE,"/properties").hasRole("INVESTOR")
+                    //.requestMatchers(HttpMethod.POST,"/tenants").hasRole("ADMIN")
+                    //.requestMatchers(HttpMethod.GET,"/tenants").hasAnyRole("ADMIN","INVESTOR")
+                    //.requestMatchers(HttpMethod.PUT,"/tenants").hasRole("TENANT")
+                    //.requestMatchers(HttpMethod.DELETE,"/tenants").hasRole("ADMIN")
+                    //.requestMatchers(HttpMethod.POST,"/investors").hasRole("ADMIN")
+                    //.requestMatchers(HttpMethod.GET,"/investors").hasRole("ADMIN")
+                    //.requestMatchers(HttpMethod.PUT,"/investors").hasRole("INVESTOR")
+                    //.requestMatchers(HttpMethod.DELETE,"/investors").hasRole("ADMIN")
+                    //.requestMatchers(HttpMethod.POST,"/leaseagreements").hasRole("INVESTOR")
+                    //.requestMatchers(HttpMethod.GET,"/leaseagreements").hasRole("INVESTOR")
+                    //.requestMatchers(HttpMethod.PUT,"/leaseagreements").hasRole("INVESTOR")
+                    //.requestMatchers(HttpMethod.DELETE,"/leaseagreements").hasRole("INVESTOR")
+
                     //.requestMatchers(HttpMethod.POST,"/authenticate").permitAll()
                     //.requestMatchers("/authenticate").permitAll()
                     //.anyRequest().denyAll()
